+++ client/javascripts/users/actions/signin-actions.js
@@ -1,5 +1,5 @@
 /**
- * authenticate actions
+ * actions for sign-in and sign-out
  */
 'use strict';
 import 'whatwg-fetch'
@@ -8,11 +8,6 @@ export const REQUEST_SIGN_IN   = 'REQUEST_SIGN_IN'
 export const RESPONSE_SIGN_IN  = 'RESPONSE_SIGN_IN'
 export const REQUEST_SIGN_OUT  = 'REQUEST_SIGN_OUT'
 export const RESPONSE_SIGN_OUT = 'RESPONSE_SIGN_OUT'
-
-export const PASSWORD_CONFIRM  = 'PASSWORD_CONFIRM'
-export const REQUEST_SIGN_UP   = 'REQUEST_SIGN_UP'
-export const RESPONSE_SIGN_UP  = 'RESPONSE_SIGN_UP'
-
 
 /**
  * action creators
@@ -39,33 +34,6 @@ export function requestSignOut() {
 export function responseSignOut(response) {
   return {
     type: RESPONSE_SIGN_OUT,
-    response
-  }
-}
-
-export function passwordConfirmFailure() {
-  return {
-    type: PASSWORD_CONFIRM,
-    error: 'CONFIRM_FAILURE',
-  }
-}
-
-export function passwordConfirmSuccess() {
-  return {
-    type: PASSWORD_CONFIRM,
-    error: 'CONFIRM_SUCCESS',
-  }
-}
-
-export function requestSignUp() {
-  return {
-    type: REQUEST_SIGN_UP,
-  }
-}
-
-export function responseSignUp(response) {
-  return {
-    type: RESPONSE_SIGN_UP,
     response
   }
 }
